pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "screener"
version = "0.1.0"
description = "Deterministic equity screening pipeline with evidence-grounded reporting"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "mcp>=1.13.0",
  "google-genai>=1.33.0",
  "httpx>=0.28.1",
  "jinja2>=3.1.4",
  "markdown>=3.7",
  "openai>=1.68.2",
  "pydantic>=2.10.6",
  "pydantic-settings>=2.7.1",
  "typer>=0.15.1",
  "weasyprint>=64.0",
]

[project.optional-dependencies]
dev = [
  "basedpyright>=1.38.2",
  "pytest>=8.3.5",
]

[project.scripts]
screener = "screener.cli:main"
screener-mcp = "screener.mcp_server:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
